requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.11
botocore==1.40.11
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
Django==5.2.5
django-allauth==0.63.6
django-cors-headers==4.1.0
django-countries==7.6.1
django-cron==0.6.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
et-xmlfile==1.1.0
Faker==37.5.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
jmespath==1.0.1
lxml==5.3.0
MouseInfo==0.1.3
numpy==2.1.0
openai==1.44.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyGetWindow==0.0.9
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.0
python-slugify==8.0.4
pytweening==1.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.32.5
s3transfer==0.13.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
text-unidecode==1.3
tk==0.1.0
tqdm==4.66.5
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
